Split 'Data' folder into 'mapbox' and 'mock_data' folders
