+++ src/mapbox/LayerStyles.js
