+++ src/mapbox/Map.js
@@ -1,0 +1,2 @@
+import React from 'react'
+
